+++ script/core/hover/return.lua
@@ -63,7 +63,10 @@ local function asFunction(source)
     for i = 1, num do
         local rtn  = vm.getReturnOfFunction(source, i)
         local doc  = docs[i]
-        local name = doc and doc.name and doc.name[1] and (doc.name[1] .. ': ')
+        local name = doc and doc.name and doc.name[1]
+        if name and name ~= '...' then
+            name = name .. ': '
+        end
         local text = rtn and ('%s%s'):format(
             name or '',
             vm.getInfer(rtn):view(guide.getUri(source))
@@ -85,6 +88,13 @@ local function asDocFunction(source)
     local returns = {}
     for i, rtn in ipairs(source.returns) do
         local rtnText = vm.getInfer(rtn):view(guide.getUri(source))
+        if rtn.name then
+            if rtn.name[1] == '...' then
+                rtnText = rtn.name[1] .. rtnText
+            else
+                rtnText = rtn.name[1] .. ': ' .. rtnText
+            end
+        end
         if i == 1 then
             returns[#returns+1] = ('  -> %s'):format(rtnText)
         else